+++ NEWS.md
@@ -1,4 +1,5 @@
 # contactsurveys (development version)
 
 * Added NEWS file
-* Added verbosity to list_surveys()
+* Added verbosity to `list_surveys()`
+* Added `get_citation()` - Resolves #38